# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/ship-order": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Ship Order Action",
                "description": "Completing order and getting shipping date",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "User Token",
                        "required": true
                    },
                    {
                        "name": "order_id",
                        "in": "query",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order shipped successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/order-detail": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Order Detail Action",
                "description": "Showing the detail of an order with the given id",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "User Token",
                        "required": true
                    },
                    {
                        "name": "order_id",
                        "in": "query",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order detail is displayed successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/all-orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "All User Orders",
                "description": "All orders that belongs to the specified user",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "User Token",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "All user orders displayed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "order_code": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "user_address": {
                                            "type": "string"
                                        },
                                        "shipping_date": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/update-order": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Update Order Action",
                "description": "Updating the product of the order that has not been shipped",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "User Token",
                        "required": true
                    },
                    {
                        "name": "order_id",
                        "in": "query",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_product_id",
                        "in": "query",
                        "description": "Order Product ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Product Quantity",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order product is updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/add-to-cart": {
            "post": {
                "tags": [
                    "Shopping Cart"
                ],
                "summary": "Add To Shopping Cart Action",
                "description": "Adding product to the shopping cart",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "User Token",
                        "required": true
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Product Quantity",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product added to the shopping cart successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/confirm-cart": {
            "post": {
                "tags": [
                    "Shopping Cart"
                ],
                "summary": "Confirm Shopping Cart Action",
                "description": "Confirming shopping cart products",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "User Token",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shopping cart is confirmed successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login Action",
                "description": "Login Action",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User e-mail address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Is Successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}